removed no longer necessary types in rate-limiter crate - helpers related with impls of AsyncRead, etc.
